+++ src/main/java/com/sparta/cucumber/chat/ChatRoomRepository.java
@@ -1,6 +1,5 @@
-package com.sparta.cucumber.repository;
+package com.sparta.cucumber.chat;
 
-import com.sparta.cucumber.models.ChatRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatRoomRepository extends JpaRepository<ChatRoom,Long> {
