+++ src/main/java/com/sparta/cucumber/chat/EnterRoomRepository.java
@@ -1,7 +1,5 @@
-package com.sparta.cucumber.repository;
+package com.sparta.cucumber.chat;
 
-import com.sparta.cucumber.models.ChatRoom;
-import com.sparta.cucumber.models.EnterRoom;
 import com.sparta.cucumber.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
