+++ src/main/java/com/sparta/cucumber/chat/Meet.java
@@ -1,5 +1,8 @@
-package com.sparta.cucumber.models;
+package com.sparta.cucumber.chat;
 
+import com.sparta.cucumber.models.Article;
+import com.sparta.cucumber.models.Timestamped;
+import com.sparta.cucumber.models.User;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
