+++ src/main/java/com/sparta/cucumber/chat/MeetRequestDto.java
@@ -1,6 +1,6 @@
-package com.sparta.cucumber.dto;
+package com.sparta.cucumber.chat;
 
-import lombok.*;
+import lombok.Data;
 
 @Data
 public class MeetRequestDto {
