+++ src/main/java/com/sparta/cucumber/chat/MeetRestController.java
@@ -1,8 +1,5 @@
-package com.sparta.cucumber.controller;
+package com.sparta.cucumber.chat;
 
-import com.sparta.cucumber.dto.MeetRequestDto;
-import com.sparta.cucumber.models.Meet;
-import com.sparta.cucumber.service.MeetService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
